First bit of the logic for the native callback interceptor
The interceptor will be responsible for overriding amx_FindPublic and
amx_Exec, allowing the plugin to handle the event prior to sending it to
Pawn. This provides us with a clean fast-path, that can be implemented
without dependencies on certain versions of the SA-MP server.

The publics available in the gamemode AMX (to be chosen with a new
native: RegisterGamemodeAMX) will be stored in two maps, {index -> name}
and {name -> index}, ensuring std::unordered_map's O(1) access based on
each data point. This might even provide a minor speedup for gamemodes
with more than a million publics :).